+++ pkg/sidecars/lbsidecar/handlers_test.go
@@ -292,4 +292,120 @@ func deleteMockEnvVars() {
 	os.Unsetenv("INSPR_SIDECAR_RANDBROKER1_ADDR")
 	os.Unsetenv("chan5_SCHEMA")
 	os.Unsetenv("chan5_RESOLVED")
+
+}
+
+func TestServer_sendRequest(t *testing.T) {
+	createMockEnvVars()
+	defer deleteMockEnvVars()
+
+	wServer := httptest.NewServer(Init().sendRequest())
+	client := http.Client{}
+
+	tests := []struct {
+		name    string
+		path    string
+		msg     string
+		route   string
+		data    string
+		wantErr bool
+	}{
+		{
+			name:    "valid route request",
+			path:    "rt1/edp1",
+			msg:     "mock_string",
+			route:   "rt1_ROUTE",
+			data:    "http://localhost:3301;edp1;edp2",
+			wantErr: false,
+		},
+		{
+			name:    "invalid route request",
+			path:    "rt1/edp1",
+			msg:     "mock_string",
+			route:   "non_ROUTE",
+			data:    "http://localhost:3301;edp1;edp2",
+			wantErr: true,
+		},
+		{
+			name:    "invalid route endpoint",
+			path:    "rt1/edp3",
+			msg:     "mock_string",
+			route:   "rt1_ROUTE",
+			data:    "http://localhost:3301;edp1;edp2",
+			wantErr: true,
+		},
+		{
+			name:    "invalid route address",
+			path:    "rt1/edp1",
+			msg:     "mock_string",
+			route:   "rt1_ROUTE",
+			data:    "localhost:3301;edp1;edp2",
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			os.Setenv(tt.route, tt.data)
+			defer os.Unsetenv(tt.route)
+
+			testServer := createMockedRoute("3301", tt.msg)
+			testServer.Start()
+			defer testServer.Close()
+
+			buf, _ := json.Marshal(tt.msg)
+			reqInfo, _ := http.NewRequest(http.MethodPost,
+				wServer.URL+"/route/"+tt.path,
+				bytes.NewBuffer(buf))
+
+			resp, err := client.Do(reqInfo)
+			if err != nil {
+				if !tt.wantErr {
+					t.Errorf("Error while doing the request: %v", err)
+				}
+				return
+			}
+
+			if tt.wantErr && (resp.StatusCode == http.StatusOK) {
+				t.Errorf("Wanted error, received 'nil'")
+				return
+			}
+
+			if !tt.wantErr && (resp.StatusCode != http.StatusOK) {
+				t.Errorf("Received status %v, wanted %v", resp.StatusCode, http.StatusOK)
+				return
+			}
+		})
+	}
+}
+
+func createMockedRoute(port string, expectedMsg interface{}) *httptest.Server {
+	listener, err := net.Listen("tcp", "localhost:"+port)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	mockServer := httptest.NewUnstartedServer(http.HandlerFunc(
+		func(w http.ResponseWriter, r *http.Request) {
+
+			var receivedData string
+			err := json.NewDecoder(r.Body).Decode(&receivedData)
+			if err != nil {
+				rest.ERROR(w, fmt.Errorf("error while reading msg body"))
+				return
+			}
+
+			if expectedMsg != receivedData {
+				rest.ERROR(w, fmt.Errorf("invalid message"))
+				return
+			}
+
+			rest.JSON(w, http.StatusOK, nil)
+		}),
+	)
+
+	mockServer.Listener.Close()
+	mockServer.Listener = listener
+
+	return mockServer
+
 }